+++ card_game/constants.py
